Hooking up new tape calculation
The previous commit didn't actually work with the rest of the system, it was just a proof of concept.  Now it works at least for testGrammarBasic.

Changes necessary:

(1a) If you create new structure, it doesn't have a tape structure, and you have to run the CalculateTapes pass again.  This is part of why it was necessary to run the original Grammar.calculateTapes so many times, because after each pass there was a lot of new structure.  I created a method `Grammar.tapify()` which is a convenience wrapper around the `CalculateTapes` pass.  (I didn't just want to reuse `.calculateTapes`, for the following reason.)  Tapify is a bit annoying to use but we can maybe simplify things later, or build it into the pass architecture somehow.

(1b) It used to be the responsibility of each pass, if it needed Grammar.tapes at all, to be sure to call Grammar.calculateTapes first.  But that's like cleaning out the drier lint from the previous user of the washer.  It's better to clean out your own lint: each pass is now responsible for making sure the resulting tree structure has a coherent tape structure.

(1c) We don't want to RE-calculate tapes that we're already sure about, but there was no way to distinguish our initial tape state (an empty TapeSet) from a grammar whose answer is actually "I have a set of zero tapes", like Epsilon.  So the default tape before calculation is now a new object, TapeUnknown.  If it's not TapeUnknown we leave it alone during calculation.

(1d) There's a new param to CalculateTapes, recalculate, in case one DOES need to recalculate the entire tree.  That's also what we use during the second pass of recalculate, when we push the results into Embeds.