Making new tapes work with more tests
(1) There's always been an ambiguity w.r.t. what "tapes" can mean for a CollectionGrammar.  Is it all the tapes, for everything in the collection?  Or is it the tapes for the selected argument?

In the past it's always been "for everything in the collection" but that can lead to some subtly incorrect results about vocab, atomicity, etc.  So now that we're being more concrete and explicit about tape names, I've gone the other way: it's only the selected one.

This means, though, that operations that need the global inventory of tapes (like vocab determination) need to be postponed until the client selects a symbol.  That's not ideal for vocabs, because that's an expensive calculation.  We need a better way to partially collect vocab, so that the additional vocab collection at staging time becomes quick again.

(2) In the absence of a selected symbol, CollectionGrammar's tapes were represented by a TapeUnknown, but this means that the whole thing gets recalculated when a symbol is first selected.  We don't want to do that.  So now it's an empty `TapeSet`, not a `TapeUnknown`.

(3) A silly bug in ValueSet (a `return` where a `continue` should have been) was preventing us from correctly flattening nested TapeSets.