Handling rename errors during tape calculation
(1) We used to handle rename/hide error messaging and recovery in the middle of the grammar passes (in `renameFix`).  That was fine because tapes were ephemeral, it didn't really matter if/when the tape structure changed.  But now that tapes are kept intact, we want to nail down the final tape structure as early as possible.  So I've moved these checks into CalculateTapes itself and deleted `renameFix`.

(2) While I was at it, I added the ability to specify expected errors in grammar tests.  We had always had some grammar tests that had errors, but we weren't able to test whether or not those errors were caught.  Grammars made from convenience functions like Lit() and Seq() don't have locations, and unlocated messages weren't sent to the devEnv.

(2a) Unlocated messages now ARE sent to the devEnv

(2b) testGenerate now accepts a numErrors param, which is the number of errors/warnings you expect the test to find.  (Unlike testing "source" grammars, these errors aren't located anywhere, so the only thing we test about them is that they exist.)