Bug in tape system
When refactoring tape calculation in another branch, I realized that tapes are subtly but deeply bugged.  It mostly doesn't happen to come up in our current source-built grammars but it's still something we should fix -- semantic inconsistencies eventually rear their ugly head elsewhere.

The issue is an erroneous conflation between "what tapes exist in the scope of a particular grammar" and "what tapes are relevant to its interpretation".  This distinction passed us by in early development because for many grammars these are the same thing.  But for a few it was unclear.  Like a Collection, are its `.tapes` the union of its symbols, epsilon, the default tape, or the selected tape?  And now that we can have local tapes with PreTape and Cursor, it raises a similar question.  A Cursor's tape is defined to come into existence in that scope, `Cursor("x", child)`, and not be the same tape as "x" outside that scope... so is "x" part of `Cursor.tapes` or not?

`Grammar.calculateTapes` calculates the former interpretation, what tapes exist in that scope, but we're sometimes using that for the second interpretation, leading to bugs where we're making decisions based on the presence/identity of tapes that don't exist or aren't the same.

I'm illustrating this bug with joins but it's not limited to it, there are a bunch of potential bugs of this class.  (Side note: this commit also threads `Expr.forward` through joins, which is necessary for this illustration.)