More cleanup
(1) Separated the rest of utils into separate files.  The remainder ended up in a file `utils/func` for want of a better name.  But they all have a commonality that they help the rest of the code have a functional style (e.g. abbreviating common types, expressing things as operations on iterators/sets/etc. rather than looping, etc.)

(2) Eliminated `treesThatGrow.ts`, it was just an experiment.

(3) Eliminated the property testing suite that we weren't using.  We might want it back one day, but a ton has changed since then and we need to rethink from scratch what we actually want to test.  A lot of the old property tests aren't actually all that interesting now.