Prelims for vocab collection
(1a) Made a VocabSet object that unions/intersects properly with "wildcard" items.  Some grammars' vocabs aren't known at the start -- Dot can be anything in the vocab, but we don't know what that is yet, and to calculate it properly that "wildness" needs to propagate through the calculation.

(1b) Made some tests for that calculation.

(2) Vocab calculation is a lot like tape calculation, and we could potentially do that in one fell swoop.  But tape calculation is quadratic to the number of symbols -- vocab calculation is quadratic to every <symbol, tape in symbol> pair.  (Meanwhile, the way we WERE doing it is pretty bad, too.  It's our worst compilation cost, if I remember correctly.  Although in practice we don't visit most <symbol,tape in symbol> pairs, each such iteration also involves a grammar traversal, and every operation on each node involves iterating through the size of the vocab.)

In prep for wrapping my head around doing the vocab calculation efficiently, I refactored the tape calculation so that it's not quadratic anymore.  (I mean, it's still worst-case quadratic, if someone were to write a purposely malicious grammar, but now we're skipping symbols that aren't going to make a difference.)