Fixed bug in new tape calc
(1) In the last commit (or so), I refactored tape calculation to be more efficient, but made a coding error that prevented it from correctly resolving (e.g.) `Seq(Embed(a), Embed(b))`!  Rather than resolving reference `$a` and then taking the result of that and resolving reference `$b`, I was still using the original when resolving reference `$b`.  This fixes that.

(2) Changing both the calculation of tapes AND their semantics at once is tricky (let alone doing vocabs at the same time), because it's not clear when a failed test is because my calculation is wrong or because the semantics has changed.

For the moment, I'm mentally backing up and targeting the old (albeit incorrect) semantics of tapes, so I can thread the new tape calc through the system using the existing battery of tests.  Then once I'm sure it's doing what it's meant to do, I'll target the new, correct semantics and see what changes.