Making new tapes work with filters
(1) I had said that {starts,ends,contains}'s tapes were default (just the union of their children) but this wasn't quite true.

Because of the needs of filter rescoping, they can also have another tape passed into the constructor.  (This lets us handle things like the rescoping of `starts(t1:h+epsilon)`.  When this gets rescoped to `t1:h+starts(epsilon)`, `starts(epsilon)` would have no tapes and thus turn into `epsilon` rather than `t1:.*`.

So we need to pass in that `t1` somehow.)  So now, calculateTapes correctly includes that extra tape.

(2) That extra tape was held in `._tapes`, but `._tapes` in general no longer exists elsewhere, and underscored members are treated as ephemeral by the nature of `.mapChildren`.  It being ephemeral doesn't break anything, but we're moving away from ephemerality in general.  Now it's the ordinary member `.extraTapes`.

(NB: When `extraTapes` is actually used in the passes, it's known to be singular, but it's inconvenient to have it be a different type than `.tapes`, so it's staying a `string[]` for now even though that will always only have one member.)

(3) `processFilters` went through the tree and errored/fixed incorrect tape structure in filters.  (E.g., it would remove `starts t1:x` if `t1` didn't exist in the grammar, or else the resulting join of the grammar and `t1:x.*` would generate infinitely.)

This is a tape structure error and those should now be done, when possible, during the `calculateTapes` pass.  We want to hammer out the tape structure as early as possible and try to avoid any large-scale changes to it later.

So this commit removes `processFilters` and puts that logic into `calculateTapes`.

(4) We had renamed `adjustFilters` to be `adjustConditions` for semantic consistency (strictly speaking the filter is the parent node, eventually a Join, and the starts/ends/etc is the condition), but "condition" is so abstract I couldn't remember what this pass actually did.  So it's back to `adjustFilters`.