Making new tapes work with SingleTape
(1) Like everything else that creates new structure, the handleSingleTapes pass had to be modified to work with new tapes, basically by calling tapify() whenever new structure is created.

Note that while SingleTape creates Renames (and Lits/Dots with new tapeNames) and thus changes the tape structure, it only does so locally.  SingleTape didn't do the default tape calculation (union of childrens' tapes), it `.tapes` are just `{tapeName}`, and the result of this pass will have that same `.tapes`.  So from the global perspective nothing has changed.

(2) While I was at it, I noticed we had no grammar tests for SingleTape, so I made some.

(3a) The error message for `Embed.tapes.length != 1` was confusing in the case `Embed.tapes.length == 0`, so I split them up into two error messages.

(3b) We were only catching that error when the child was an Embed, it was actually fine to say `SingleTape(Seq(t1("a"),t2("b")))`.  This never actually comes up in real source grammars, but it's still erroneous and should receive an error/fixup.  So now we're checking for it.  (No error when it's `tapes.length == 0`, though, that's not problematic.)