Fixes and improvements to tests
(1) I noticed that the NotMatch tests were still using the old version of generateOutputs and had broken.  What they were doing with it is kind of a workaround, though -- we were using a grammar to generate the expected outputs to test another grammar against.

What we really needed here was a simple test function that asserts that two grammars are equal.  And we'll need that anyway in the near future for LTR/RTL and atomicity testing.  So I wrote one, testGrammarEqual(), which takes additional `grammar2` and `opt2` params and asserts the results are identical.

If you leave out `opt2` it just uses the original `opt` param (handy for equivalent testing where we are only using the first op), and if you leave out `grammar2` it just uses the original `grammar` param (handy for testing options like LTR/RTL).

(2) This necessitated some refactors elsewhere in the testing system, but for the most part the public API for the test suites remain the same.  Also these refactors caught a few problems in threading Options through the system, like where it wasn't getting to where it needed to, or was threaded to places it wasn't needed.