Tapes for Replace, SingleTape bugfix
(1) Threading the new tape system through Replace was pretty simple, but unearthed a bug in SimpleTape handling.

(2) I mentioned earlier today that `handleSingleTapes` didn't change the tape information at the global level.  But it does change the tape information at the local level, by the way it recurses through its children (seqs, alts, etc.) and potentially changes the tape of its lits.

But it wasn't changing the tapes of the seqs/alts/etc. in turn, it was just letting them be the original tapes.  So we have to recalculate them.

We don't just want to calculate this with `new CalculateTapes(recalc: true).go()`, that would recalculate everything underneath the change recursively, and this happens post-order, that would be wasteful.  We just want to update that one node with the new info from its children.  We can do that by setting the children to TapeUnknown and then calling `CalculateTapes` (via `tapify`) without a recalc.

This feels tedious and hacky, at some point we should factor this out into something more automated... but also tapify itself is a little tedious and error-prone.  I'm going to hold off until I understand our overall taping/vocab needs before automating further.

(3) It's just by accident that a Replace test happened to have the erroneous configuration, so I wrote some dedicated tests for it in `testGrammarSingleTape`.

(4) Putting vocabs inside Replace params no longer works, so one of the tests (something like 34a-alt if I remember correctly) breaks.  It's just the alternate of a more-normally-vocabbed test so I deleted it.