Fixing several bugs in filter tape calc
(1a) Trying to run the tests on testGrammarFilters exposed a bug -- in fact, our tape calc wasn't running on Filters at all!

When attempting to actually run the code, it turns out that the Filter tape-sanity checking was still written as if it's a later pass.  It's using `.tapes`... which calls `tapeToLit()`... which assumes tapes have already been resolved or else it throws an exception.

Fixing this required a new function in `tapes.ts`, which gives the length of a TapeID (in terms of number of literals) or "unknown" if there is unresolved material.

(1b) Why weren't we running that code before?  It turns out that starts/ends/contains used a different structure between grammar tests and source tests.  In grammar tests they were joined using a JoinGrammar, in source tests they used a FilterGrammar (which is mostly there for sanity checks)

We hadn't noticed the discrepancy because we didn't pay attention to grammar test errors until recently.

(1c) Also arising: the question of when to turn FilterGrammar into JoinGrammar.  There used to be a pass that both did some sanity checks and returns a JoinGrammar, but I eliminated that and moved the code into CalculateTapes.

This was a mistake!  CalculateTapes effectively runs twice, once to get the preliminary tape structure (including references), and once to finalize the tape structure with all references resolved.  But the first time it runs, it can't do these checks because the tape structure isn't finalized... and then it replaces itself with a JoinGrammar so this code never runs the second time to do those checks for real.

So I moved the JoinGrammar construction to `adjustFilters.ts`, and while I was at it renamed that to `createFilters.ts` because it's doing more than just adjusting.

(2) In source tests, we were incorrectly counting errors twice, and getting the number wrong the second time.  This required some further refactoring of the test utility functions to make sure errors were counted once and only once for both grammars and sources.