Finishing up tape calc
(1) When I rewrote the tape resolution, I had a more complicated algorithm coded, but then second guessed myself.  "Does it need to be this complicated?  I can't seem to write a test that gets it wrong..." and then replaced it with a much shorter version.

Turns out nope, we needed the more complicated version; one of the tests in `testSourceMulti` had a recursive import structure that the simpler version gets wrong.

(2) Small adjustment to `testLength` so that it calls `tapify` rather than `calculateTapes`.

(3) This takes us back to green, except for those new Cursor tests merged from the main branch (that illustrate the deeper semantic problem with the old tapes).