Symbol selection as a Pass
(1) I was wanting to simplify the Pass hierarchy a bit, and almost no one was actually using `Pass.transformRoot()`.  The only remaining use of this was in `QualifyNames`, and that code was actually really weird.

It was throwing out the results of the actual transform, and then making a new CollectionGrammar from the accumulated symbols.  Which is fine, but also, we had just a moment before returned a CollectionGrammar with no symbols in order to throw it out.  How about we just return a CollectionGrammar with the accumulated symbols instead?

The issue is that this process doesn't always result in the root being a CollectionGrammar -- sometimes it's a LocationGrammar.  The weird old code DOES always result in it being a CollectionGrammar.  And things break when the root isn't a CollectionGrammar.

But what things and why?  Turns out there's only actually one thing left that cares: we call `selectSymbol` on the root grammar, and if this isn't a CollectionGrammar, oops failure.  This whole chain of weirdness and brittleness is only supporting one operation.  If symbol selection were actually a recursive algorithm instead, it wouldn't care that the root is a LocationGrammar.

(2) So let's do that: symbol selection is now a Pass rather than a function.  It's not in ALL_PASSES, though, because those happen before we know what symbol the client actually wants.  (This is the first Pass in a new, later stage of Passes, which I usually call "Staging", as in staging a grammar to respond to a particular query.)

(3) This exposed some subtle issues in tapePriority and determineAtomicity.  Rather than get the symbols from a CollectionGrammar pushing them into the `env`, we were pre-pushing them into the `env` ourselves and the tapePriority was using those.  (That's not the best -- several things happen in the meantime, what if one of them were to change that?  Nothing does that NOW, but it's a brittle assumption to make.)

Then, neither of them was actually using the priority/atomicity of the selected symbol, they were considering it over the whole Collection, which is a different answer.  (It's similar to our problem with new/old tape semantics, but we shouldn't even have been doing this in the old semantics.)  So I fixed that, too.

(Back to 1) So all together, this eliminates the need to make the root a CollectionGrammar. (NB: in practice it still has to be close to the root, I didn't let symbol selection recurse anything but a LocationGrammar.  If you nest something else around the Collection, make sure you change SelectSymbol to recurse through it.)  With that need, we're able to lose the weird code in `QualifyNames.transformRoot()`, which lets us eliminate `Pass.transformRoot()` altogether.