Fixed bug in infinityProtection
(1) The previous commit found a bug in several places where CollectionGrammar wasn't pushing its symbols into `env`, and also not properly using its selected symbol.

Looking around for other places this was happening, I found that the length estimate in infinityProtection was doing the same thing.  Or rather, it was achieving nothing at all.  CollectionGrammar was being treated as an epsilon whether or not it had a symbol selected, so it was saying "my length is zero" regardless.  And since everything is a CollectionGrammar at/near the root, everything was finite, therefore infinityProtection wasn't happening.  Oooops.  This commit fixes that.

(2) I also made some very simple tests showing infinityProtection working, so that we have a warning if it stops working again.  At some point we should merge these with testLength, which is similar in its goal.

Why weren't we testing for this before?  Well, until recently we couldn't pass arbitrary Options to grammar tests, we could only pass verbosity.  I wasn't about to make a test with 100 character outputs, nor change DEFAULT_MAX_CHARS every time.

(3) The use of maxChars inside infinityProtection still adjusted for the ancient off-by-one error in token counting.  Anyway that's gone now, may we never find another.