Moving responsibility for name resolution
A while back, we hit a little snag where client requests for symbols from Interpreter were using a different algorithm for symbol name resolution than the original QualifyNames, leading to potential differences of opinion.  The issue was that, by the time the client asks for symbols, the name structure of the original grammar has already disappeared during QualifyNames.  The quick-n-easy solution to this was just for Interpreter to keep a reference to the old, pre-QualifyNames version of its grammar.

But it's a bit awkward, because we need name resolution to get the right name to ask SelectSymbol for, but name resolution depends on a weird part of Interpreter's state.  It'd be better to be able to read this right off the grammar.

So this commit moves the responsibility for name resolution to the grammar itself.  Rather than keep around a reference to the actual old grammar, CollectionGrammar has a new member "resolver" that replicates the name structure of the original grammar in a minimal form, and suffices to qualify names.  When the Interpreter needs to qualify a name or retrieve a symbol referent for other reasons (like a tape inquiry), it can access that functionality by recursing briefly into the grammar with `resolveName`.

Making this functionality recursive also removes the old requirement that `Interpreter.grammar` be a CollectionGrammar specifically.  We're no longer calling any CollectionGrammar-specific functions from Interpreter.  (Again in practice, we only recurse through LocationGrammars to find the CollectionGrammar within, so if you want to wrap some other grammar around it too, then make sure that the `resolveNames`, `SelectSymbol` etc. algorithms recurse through that kind of grammar.)