Cursors and infinity protection now Passes
(1) Adding Cursors according to the tape priority, and adding Counts according to length estimation, are now Passes.  This should have been easy but led to two bugs that took me way too long to figure out.

(2) Note adding Cursors and Counts now happen in the opposite order.  It used to be that we determined tape priority outside of adding Cursors/Counts, and then passed that into the cursor/count adding functions so they knew which tapes to operate on.  We don't want to do that anymore (we want to pass in as little information to a Pass as possible).

Instead, the new `CreateCursors` pass calculates priority on its own and adds Cursors accordingly, and then `InfinityProtection` acts specifically on Cursors, adding counts underneath them.

(Since `InfinityProtection` is recursive, it can potentially even reach Cursors that already existed in the grammar.  However, we aren't actually recursing into Collections here so we're not reaching them.  We can think about that later.)

(3) Baffling bug #1 led to infinite generation... only when adding infinity protection to a tape!!!  Why?  Well, it turned out we never threaded `Expr.forward()` through CursorExpr, and so a Cursor nested inside a Count will never do its thing.  Meanwhile, particular configurations of Match will lead a tape to be judged finite, because it's matched to a second tape that is finite, but if we never actually execute that second tape there's nothing stopping the first tape from generating forever.

Let's keep this in mind if we ever do a special Cursor execution order where we finish off a whole tape first before moving onto the next.  Matches mean that whether or not a tape is infinite depends on tape order!  This *should* be okay in real source grammars because their order is constrained by enclosing PreTapes but we should be careful when calculating it in general, we can't just use our existing tape length calculation.

(4) Baffling bug #2 was leading to infinite generation whenever a grammar had several infinite repetitions, and InfinityProtection wasn't catching this.

Why not?  JAVASCRIPT WEIRDNESS.  What's the length of `t1:o*` with respect to tape t2?  Well, `t1:o` is length zero with respect to t2, and that times infinity is also zero, right?

Nope, `0 * Infinity == NaN`.  (It's complicated why.)  This NaN then propagates through the system, and in some configurations can lead to a tape having NaN length overall, which is not equal to Infinity, so InfinityProtection wasn't being applied.

So now we don't use plain JavaScript multiplication, we just make a little auxiliary multiplication function that gives us the zero we want.