Small changes
(1) Sometimes we called the query "query" and other times "restriction".  It's always "query" now.

(2) DerivEnv was keeping a `Namespace<Expr>`, but we don't actually use the stack-like properties of Namespaces, we only ever use it like a Dict.  So now it's a `Dict<Expr>`.  If we ever actually need a stack, we can change it back.

(3) Now if you pass a bad string to getTapeColor it throws rather than ignores it.