Figuring out the types of complex passes
In the big picture, our grammar is a tree that represents a graph (because of Embeds & Collections).  Our algorithms usually do not have to engage with the graph, though, most are tree-to-tree transformations.  When they *do* need to engage with the graph-ness, though, they use a variety of actual algorithms to do so.  It'd be better to work out one Pass-like algorithm they all share...

But not all the graph algorithms we need are actually Component => Component functions.  What about Component => X transformations?  For example, the length functions are `Grammar => LengthRange`; atomicity is `Grammar => { joinable: boolean, concatenable: boolean }`.

All of these handle their own recursion in an ad-hoc manner, and one reason for this is that "recursion strategy" helper functions like `mapChildren` becomes difficult to actually express in Typescript.  I'd never been able to figure out what the type of mapChildren would be, in cases where it actually changed types (like TSTs => Grammars or Grammars => Exprs).  Everything I tried led to typechecking errors in one place or another.

I eventually figured it out.  It requires something I call "Skeleton", an object that has all the keys of X, but whenever the corresponding value of a key is a Component, it's now type Y instead.  For example, the number skeleton of JoinGrammar looks like a JoinGrammar -- its tag is still "join", it has tapes, etc.  -- but `.child1` and `.child2` are numbers instead of Grammars.

It's monstrous to actually express, and we don't need it for anything at this moment , but I'm committing it so that the commit history has the answer to a longstanding question.