Full property testing framework
I wanted to check whether the tape calc algorithm was order-independent: does it matter in which order we calculate symbols?  We don't want it to be!  The proptest framework found examples where it mattered, but they were too complex for a human to work out what was going wrong.

Previously, we had only ever done the first half of property testing: finding random failing examples.  If we have some function `increment(x)` and our proptest framework randomly tries it on 13408509 and something fails, that's not super helpful.

What's more helpful is the answer to the question "What's the LOWEST number on which it fails?"  If it fails on 32768 but succeeds on 32767, that's more helpful information: it tells us that signed 16 integers are probably to blame.

So proptest frameworks have a second step after generation, where they try to reduce the failing example until they find a minimal version of it that still fails.

This is harder for grammars than integers, obviously; we can't practically do a full search for the most minimal example.  We're effectively doing a random greedy search for simpler examples:

(1) Take the most-recently-found failing example
  (2) Randomly reduce it (turn a literal to an epsilon, drop a symbol, etc.)
    (3a) If the reduced example fails, that's our new minimal failure, go back to step 1
    (3b) If the reduced example succeeds, go back to step 2 and generate another
  (4) If we've generated 100 reduced examples and all of them succeed, then the most-recently-found failure is probably close to minimal, return it.

This actually works quite well; the minimal failures for the above issue make it pretty obvious what's going wrong.  They're all cases where there's a mutually recursive reference and one of them is inside a rename.

This is a paradox, it's no surprise that the tape calc is getting hung up on this.  It'd be fine if the checker just gave up here -- hopefully by making everything as right as possible and issuing a comprehensible error, of course.  But what's actually happening is that our ability to detect that something is going wrong is dependent on the order in which the mutually-recursive symbols are evaluated!