Refactoring tape objects
The VocabInfo objects were inspired by TapeInfo (then: "TapeID") objects, but the vocabs did them one better.

The tape objects had singular literals (TapeLit) containing only one tape.  To have multiple, you needed a TapeSet object. But the TapeSet object could contain either literal or unresolved (ref, rename, and unknown) content.  This made several operations awkward because they had to recurse into the TapeSets to know whether they were entirely literal or partially unresolved.  (E.g., you didn't know from something being a TapeSet whether it has a length, so there had to be a special recursive tapeLength function, and it had to be able to return an "unknown" result if there was unresolved material.)

Vocabs didn't have this problem.  The basic objects are all "set of literals".  If any operation (e.g. summing) happens where all the operands are literals, the operation happens immediately and results in another set-of-literals object.  If any of the operands are not sets-of-literals, the result is an unfinished calculation object like VocabSum.  This means the literalness of an object is 100% clear from its type -- you don't have to introspect into a VocabSum to know that it's got unresolved content in it, because if it didn't it wouldn't be a VocabSum in the first place.

So this commit does the same for TapeInfo types.  TapeLit is now the set object, and the only thing it contains are literals.  There's no TapeSet object anymore, but there's an object representing an unfinished sum operation, TapeSum.

This gets rid of some awkward functions like tapeLength, and doing things like calling tapeLength when what we really want to know is "is this literal?"  Now it's obvious what's literal, it's literal if and only if it's a TapeLit.