Eliminating LocatorGrammar
I've made some good progress figuring out a TapeCalc algorithm that can catch impossible rename-embed paradoxes, but while I'm pondering I'm doing some refactoring to handle an annoying prerequisite for this change: "How do we know where to send the associated error messages?"

The errors are caught in the tape resolution stage, but we can't actually pass locations into these objects/operations, not easily... because Grammars don't know their locations, and haven't for a long time.  Instead, the Grammar tree is associated with locations indirectly, through a decorator class called LocatorGrammar.

Locators had their pluses and minuses.  On the plus side:

  (1) Grammars not having positions saved a lot of typing (e.g. passing `pos` params into every constructor) and error-prone boilerplate.
  (2) Sometimes it was rather hard to associate grammars with particular positions.  So long as these grammars are eventually inside a LocatorGrammar, their messages will be localized to a reasonable position.
  (3) I also thought this would let us do hierarchically-marked error messages in case of row-folding. (But the way the locator system actually evolved precludes this, we'd have to do this a different way anyway.)

On the minus side:

  (4) Locator grammars really mess with `instanceof` and pattern-matching.  You can't just say "If child is a Collection" because it might be a Collection wrapped in a Locator.
  (5) And as mentioned, it makes it hard to pass position information to further-downstream error-checking systems.

This update keeps the basic idea and mechanics of LocatorGrammars (especially the ability to do (2)), but takes away the actual object that made (4) so annoying.  Now any Grammar can have a `pos`, and if it's set then that grammar acts like a LocatorGrammar.

Unlike before, though, (1) is not such a huge issue anymore, because of `mapChildren` and `update` making it easier to just pass members between generations of Grammars rather than have to explicitly pass a `pos` into constructors, and because of convenience passes like `PostPass` auto-localizing error messages.

We could probably automate this further, like this did require some annoyingly manual position juggling, but that's something for another day.  For now, the goal is just that (most) Grammars know their position again and can pass this on to other systems if/when necessary.