New proptest reveals Join-Embed-Rename tape bug
(1) I resurrected another part of the old Property Test suite, in which we could express equations like `X|Y == Y|X` and assert them to be true of random grammars X and Y.

Last time, this was expressed by passing in ad-hoc functions for the left and right sides, like the "left" param might be:

    `(x: Grammar, y: Grammar) => Join(x,y)`

This time, you pass in what symbols you want to exist (e.g. `["X", "Y"]`) and then just pass in a Grammar that embeds them:

    `Join(Embed("X"), Embed("Y"))`.

This lets us utilize some of the existing machinery of Collections rather than write it from scratch.

(2) The above test revealed a bug in how we detect and recover from rename errors.  The configuration required to show it is strange and I wouldn't have come up with it myself, but the bug is pretty general.

The issue seems to be, in the second pass of tapecalc, we can catch erroneous renames that we couldn't originally... but by the time we catch them and create an `.ERR` tape to fix things, we've already set in stone other tape decisions that depend on this.

We need to be catching this error earlier.  We couldn't before, actually, because we didn't have error-messaging capability during tape unification.  The previous commit makes that possible, however.