Determining scope of rename bug
(1) The previous bug doesn't have anything specific to do with Joins, it also happens with Unions.  (That's reassuring, actually; if it was Joins specifically then we have to worry about the tape-knowledge the join has.)

I made Union versions of the tests and moved them to testGrammarRename, because they're really Rename bugs not Join bugs.

(2) While I was at it, I filled out a bunch more equation proptests.