Fixing recursive rename bugs
A few relatively minor changes fixed our various bugs about tape calculation, renaming, and recursion.  (The code has been somewhat refactored because I expected more serious changes, but my fixes here probably would have worked with our old code too.)

(1) I mentioned in a meeting that I had made an error in conflating two meanings of TapeInfo -- one (call it "intensional") that it represents a constraint on what tapes the symbol can contain, and two (call it "extensional") that it represents that set of tapes.

All along, we've been saying "If you encounter a recursive reference to a symbol, don't just recurse infinitely, just return the empty set."  From an extensional point of view, that's stupid, we know the symbol doesn't contain the empty set.  It makes sense intensionally, though -- recursing wouldn't put any more constraint on the symbol than it already has.

Using an extensional meaning was causing some problems with Renames.  In particular, if there was an erroneous rename to an existing tape, the Renames were *usually* catching this but *sometimes* failing to catch this in a recursive context.  The reason is because they were working on the extension -- looking at that recursion-breaker empty set, saying "Well THIS doesn't contain the problematic tape", and failing to complain, and failing to put the output tape into the tapes object because the input tape wasn't present either.

Part of the fix is that the Rename has to put its toTape into the tapes no matter what -- because the *symbol* will contain the input tape even though that *tape set* is empty.

So long as this happens, the "oops renaming to an already-existing tape" error reliably gets caught in the tape-pushing stage that follows.

(2) I mentioned a few commits ago that we had to do some error detection and repair during unification itself.  I even refactored the locator system to allow that.  Turns out I was wrong, we still can't do it then.  (I'm still happy I refactored the locator system, though, that's good anyway.)

We might have been able to detect and message rename errors during unification, but we couldn't FIX them.  Fixing means changing the grammar tree, and we're not in the grammar tree right now, we're playing with a little reduced form of the grammar.  So either (a) we go back to unifying on the tree itself so we can change it, or (b) we accept that errors can only be caught/fixed in the tree traversal.

(3) Turns out we probably have to accept (b) anyway.  Trying to do a graph algorithm on a graph *that might be changing out from under you* would be a nightmare.

Instead, we come to the best conclusion we can, push it to the graph, and then if the tree changes due to errors, we take the changed tree and do the whole infer-unify-push process again.  Yes, that's a tough bullet to bite but it's inevitable -- a different tree means different tapes.  Our tape calc ordering problems were in part due to using "stale" tape information after the tree had already changed.

(4) So we've addressed a lot of the recursive renaming issues, to the point that rename errors are getting caught and the repairs are happening.

Turns out our repairs are a problem too!

Backing up, the problem with renaming-to-an-existing tape (say, a `t1->t2` rename where `t2` already exists on the child) is that any `t2` material inside the scope of that rename never gets handled during the generation loop.  The rename gets a query on `t2`, and forwards it on as a `t1`; there's never any point where the material inside receives a query.  So our solution was to sneak in an additional rename inside the erroneous rename, renaming the old inaccessible `t2` material to `.ERRt2`.

However, consider a recursive rename paradox with this issue.  We encounter a paradoxical rename, and it's caught in the push stage, and we slip in that sneaky rename to `.ERRt2` to fix it.  Now we restart the tape calc process, because the tree has changed and our old answers are stale.

We recalculate, push the answers into the tree... and find a new error.  Y'see, we find this rename to a tape `.ERRt2`... but `.ERRt2` already exists in its scope due to the recursive reference.  We'll never manage to reach the `.ERRt2` material inside!  In order to reach it, we have to slip in a sneaky rename to a new tape; it gets called `.ERR.ERRt2`.  Then we restart, recalculate, push the answers into the tree, and find a problem with `.ERR.ERR.t2`...

In a way, this is correct behavior given the paradox!  Those really aren't all the same tape, each recursion does really need a different error tape to catch that material.  But it leads to an infinite loop that is tricky to actually stop, and doesn't actually *resolve* the paradoxical situation, like it's treating the symptom but not the disease.

Really, the way to remove the paradox is the simpler repair, what we do for other rename issues: drop the rename entirely.